chore(core): narrow PeerId to u8, add borsh + redb key, drop dead threshold_crypto encoding
- PeerId(u16) -> PeerId(u8). 256 guardians is well above any practical
  federation size; u8 collapses key bytes and integer-ordered range
  scans stay byte-lex-correct. Updates every caller site.
- Derive BorshSerialize/BorshDeserialize on PeerId and install a direct
  redb::Key impl via redb_newtype_key!(PeerId, u8). First core type
  carrying the Phase 3 "direct" pattern beyond the eventlog spike.
- Delete fedimint-core/src/encoding/threshold_crypto.rs — the
  Encodable/Decodable impls for PublicKeySet/PublicKey/Ciphertext/
  DecryptionShare are unused in v2 (v2 uses tbs/tpe crates, not
  threshold_crypto). Config still imports G1/G2 Projective for DKG
  internals; that lives elsewhere.

Incidental: rustfmt collapsed a few multi-line signatures in
fedimint-redb/src/lib.rs.

+++ fedimint-server/src/config/setup.rs
@@ -315,11 +315,11 @@ impl ISetupApi for SetupApi {
             .expect("We inserted the key above.");
 
         let params = ConfigGenParams {
-            identity: PeerId::from(our_id as u16),
+            identity: PeerId::from(our_id as u8),
             iroh_api_sk: local_params.iroh_api_sk,
             iroh_p2p_sk: local_params.iroh_p2p_sk,
             peers: (0..)
-                .map(|i| PeerId::from(i as u16))
+                .map(|i| PeerId::from(i as u8))
                 .zip(state.setup_codes.clone().into_iter())
                 .collect(),
             meta: BTreeMap::from_iter(vec![(
